feat(settings): Weather tab — ActiveSky master switch + weather announce toggles
New WeatherPanel (second tab): the ActiveSkyEnabled switch plus the four
weather auto-announcement controls moved verbatim from AnnouncementsPanel
(same setting keys, no persistence migration). AnnouncementsPanel keeps
only its General group.
