+++ MSFSBlindAssist/Forms/Settings/AnnouncementsPanel.cs
@@ -4,8 +4,10 @@ using MSFSBlindAssist.Settings;
 namespace MSFSBlindAssist.Forms.Settings;
 
 /// <summary>Announcements section of the unified Settings dialog. Extracted from the retired
-/// standalone announcement-settings dialog — same controls, same AccessibleNames, but the old
-/// General/Weather inner TabControl is replaced with two stacked GroupBoxes.</summary>
+/// standalone announcement-settings dialog — same controls, same AccessibleNames. The old
+/// General/Weather inner TabControl was first replaced with two stacked GroupBoxes; the
+/// Weather group has since moved to its own top-level Settings tab (WeatherPanel), leaving
+/// this panel with only the General group.</summary>
 public class AnnouncementsPanel : UserControl, ISettingsPanel
 {
     // ── General group ────────────────────────────────────────────────────────
@@ -15,16 +17,6 @@ public class AnnouncementsPanel : UserControl, ISettingsPanel
     private ComboBox _nearestCityIntervalCombo = null!;
     private CheckBox _timeWithSecondsCheck = null!;
     private CheckBox _gsxBackgroundMonitoring = null!;
-
-    // ── Weather group ────────────────────────────────────────────────────────
-    private CheckBox _weatherAutoAnnounce = null!;
-    private ComboBox _weatherIntervalCombo = null!;
-    private CheckBox _sigmetAlerts = null!;
-    private CheckBox _pirepAlerts = null!;
-    private NumericUpDown _proximityRange = null!;
-
-    /// <summary>Combo entries: minutes (0 = AS download interval, no extra throttle).</summary>
-    private static readonly int[] IntervalChoicesMinutes = { 0, 5, 10, 15, 20, 30, 45, 60 };
 
     public string TabTitle => "Announcements";
 
@@ -39,11 +31,8 @@ public class AnnouncementsPanel : UserControl, ISettingsPanel
         AutoScroll = true;
 
         var generalGroup = BuildGeneralGroup();
-        var weatherGroup = BuildWeatherGroup();
-        weatherGroup.Location = new System.Drawing.Point(12, generalGroup.Bottom + 12);
 
         Controls.Add(generalGroup);
-        Controls.Add(weatherGroup);
     }
 
     private GroupBox BuildGeneralGroup()
@@ -151,100 +140,6 @@ public class AnnouncementsPanel : UserControl, ISettingsPanel
         return group;
     }
 
-    private GroupBox BuildWeatherGroup()
-    {
-        var group = new GroupBox
-        {
-            Text = "Weather",
-            Location = new System.Drawing.Point(12, 0),
-            Size = new System.Drawing.Size(460, 230),
-            AccessibleName = "Weather",
-            AccessibleDescription = "Weather and advisory auto-announcement settings",
-        };
-
-        _weatherAutoAnnounce = new CheckBox
-        {
-            Text = "Auto-announce &weather state changes",
-            Location = new System.Drawing.Point(12, 24),
-            Size = new System.Drawing.Size(420, 24),
-            AccessibleName = "Auto-announce weather state changes",
-            AccessibleDescription = "Automatically announce when entering or leaving clouds and when precipitation starts or stops"
-        };
-
-        _sigmetAlerts = new CheckBox
-        {
-            Text = "Auto-announce approaching &SIGMETs and AIRMETs",
-            Location = new System.Drawing.Point(12, 60),
-            Size = new System.Drawing.Size(420, 24),
-            AccessibleName = "Auto-announce approaching SIGMETs and AIRMETs",
-            AccessibleDescription = "Announce when the aircraft enters the proximity range of an active SIGMET or AIRMET"
-        };
-
-        _pirepAlerts = new CheckBox
-        {
-            Text = "Auto-announce approaching pilot reports (&PIREPs)",
-            Location = new System.Drawing.Point(12, 96),
-            Size = new System.Drawing.Size(420, 24),
-            AccessibleName = "Auto-announce approaching PIREPs",
-            AccessibleDescription = "Announce when the aircraft enters the proximity range of a significant pilot report of turbulence or icing"
-        };
-
-        var rangeLabel = new Label
-        {
-            Text = "&Proximity range (nautical miles):",
-            Location = new System.Drawing.Point(12, 138),
-            Size = new System.Drawing.Size(250, 20),
-            AccessibleName = "Proximity range label"
-        };
-
-        _proximityRange = new NumericUpDown
-        {
-            Location = new System.Drawing.Point(270, 134),
-            Size = new System.Drawing.Size(80, 24),
-            Minimum = 10,
-            Maximum = 500,
-            AccessibleName = "Proximity range in nautical miles",
-            AccessibleDescription = "Distance at which to announce approaching SIGMETs, AIRMETs, and PIREPs"
-        };
-
-        var intervalLabel = new Label
-        {
-            Text = "Weather announcement &interval:",
-            Location = new System.Drawing.Point(12, 178),
-            Size = new System.Drawing.Size(250, 20),
-            AccessibleName = "Weather announcement interval label"
-        };
-
-        _weatherIntervalCombo = new ComboBox
-        {
-            Location = new System.Drawing.Point(12, 202),
-            Size = new System.Drawing.Size(338, 24),
-            DropDownStyle = ComboBoxStyle.DropDownList,
-            AccessibleName = "Weather announcement interval",
-            AccessibleDescription = "Minimum minutes between auto-announced ActiveSky weather updates. Active Sky download interval means no extra throttle; the announcer follows ActiveSky's own refresh cadence."
-        };
-        foreach (int minutes in IntervalChoicesMinutes)
-        {
-            _weatherIntervalCombo.Items.Add(IntervalChoiceLabel(minutes));
-        }
-
-        group.Controls.AddRange(new Control[]
-        {
-            _weatherAutoAnnounce, _sigmetAlerts, _pirepAlerts,
-            rangeLabel, _proximityRange,
-            intervalLabel, _weatherIntervalCombo
-        });
-
-        return group;
-    }
-
-    private static string IntervalChoiceLabel(int minutes)
-        => minutes == 0
-            ? "Active Sky download interval"
-            : minutes == 60
-                ? "1 hour"
-                : $"{minutes} minutes";
-
     private void UpdateScreenReaderStatus()
     {
         try
@@ -312,12 +207,6 @@ public class AnnouncementsPanel : UserControl, ISettingsPanel
         _nearestCityIntervalCombo.SelectedIndex = IntervalToIndex(settings.NearestCityAnnouncementInterval);
         _timeWithSecondsCheck.Checked = settings.AnnounceTimeWithSeconds;
         _gsxBackgroundMonitoring.Checked = settings.GsxBackgroundMonitoring;
-
-        _weatherAutoAnnounce.Checked = settings.WeatherAutoAnnounceEnabled;
-        _sigmetAlerts.Checked = settings.SigmetProximityAlertsEnabled;
-        _pirepAlerts.Checked = settings.PirepProximityAlertsEnabled;
-        _proximityRange.Value = Math.Clamp(settings.SigmetProximityRangeNm, 10, 500);
-        _weatherIntervalCombo.SelectedIndex = Math.Max(0, Array.IndexOf(IntervalChoicesMinutes, settings.WeatherAutoAnnounceIntervalMinutes));
     }
 
     public bool Validate(out string error, out Control? focus)
@@ -335,13 +224,6 @@ public class AnnouncementsPanel : UserControl, ISettingsPanel
         settings.NearestCityAnnouncementInterval = IndexToInterval(_nearestCityIntervalCombo.SelectedIndex);
         settings.AnnounceTimeWithSeconds = _timeWithSecondsCheck.Checked;
         settings.GsxBackgroundMonitoring = _gsxBackgroundMonitoring.Checked;
-
-        settings.WeatherAutoAnnounceEnabled = _weatherAutoAnnounce.Checked;
-        settings.WeatherAutoAnnounceIntervalMinutes = IntervalChoicesMinutes[
-            Math.Clamp(_weatherIntervalCombo.SelectedIndex, 0, IntervalChoicesMinutes.Length - 1)];
-        settings.SigmetProximityAlertsEnabled = _sigmetAlerts.Checked;
-        settings.PirepProximityAlertsEnabled = _pirepAlerts.Checked;
-        settings.SigmetProximityRangeNm = (int)_proximityRange.Value;
     }
 
     public void OnLeaving()
